# [win-perf] reduce number of powershell commands