@chris-rock 2017-10-17
@sdelano This is so cool to see the Windows performance to be improved in that magnitude. Thank you for showcasing this! 

I was thinking if we should go a route with caching of a identical shell call-outs in train instead having class variables. The reason behind my thinking is the challenge in ensuring that users do not use the same resource with different arguments. What do you think?