@adamleff 2017-12-05
Closing in favor of #2309 and chef/train#214. Thanks for your efforts, @sdelano!